move paramenters to a different locations
