+++ uyuniadm/cmd/cmd.go
@@ -22,27 +22,12 @@ func NewUyuniadmCommand() *cobra.Command {
 	rootCmd.PersistentFlags().StringVarP(&globalFlags.ConfigPath, "config", "c", "", "configuration file path")
 
 	migrateCmd := migrate.NewCommand(globalFlags)
-	addCommonFlags(migrateCmd)
 	rootCmd.AddCommand(migrateCmd)
 
 	installCmd := install.NewCommand(globalFlags)
-	addCommonFlags(installCmd)
 	rootCmd.AddCommand(installCmd)
 
 	rootCmd.AddCommand(uninstall.NewCommand(globalFlags))
 
 	return rootCmd
-}
-
-func addCommonFlags(cmd *cobra.Command) {
-	cmd.Flags().StringSlice("podman-arg", []string{}, "Extra arguments to pass to podman, separated by commas")
-
-	cmd.Flags().String("helm-uyuni-namespace", "default", "Kubernetes namespace where to install uyuni")
-	cmd.Flags().String("helm-uyuni-chart", "oci://registry.opensuse.org/uyuni/server", "URL to the uyuni helm chart")
-	cmd.Flags().String("helm-uyuni-version", "", "Version of the uyuni helm chart")
-	cmd.Flags().String("helm-uyuni-values", "", "Path to a values YAML file to use for Uyuni helm install")
-	cmd.Flags().String("helm-certmanager-namespace", "cert-manager", "Kubernetes namespace where to install cert-manager")
-	cmd.Flags().String("helm-certmanager-chart", "", "URL to the cert-manager helm chart. To be used for offline installations")
-	cmd.Flags().String("helm-certmanager-version", "", "Version of the cert-manager helm chart")
-	cmd.Flags().String("helm-certmanager-values", "", "Path to a values YAML file to use for cert-manager helm install")
 }
