+++ uyuniadm/cmd/uninstall/podman.go
@@ -6,9 +6,9 @@ import (
 	"os"
 	"os/exec"
 
-	"github.com/uyuni-project/uyuni-tools/shared/podman"
 	"github.com/uyuni-project/uyuni-tools/shared/types"
 	"github.com/uyuni-project/uyuni-tools/shared/utils"
+	"github.com/uyuni-project/uyuni-tools/uyuniadm/shared/podman"
 )
 
 func uninstallForPodman(globalFlags *types.GlobalFlags, dryRun bool, purge bool) {
